requirements.txt
pyqt5
pystray
Pillow
